`set` - refactored to use standard subcommand structure
- also removed static data as a byproduct!
