+++ commands/set/subcommands/index.js
@@ -1,0 +1,15 @@
+module.exports = [
+	require("./ambassador.js"),
+	require("./birthday.js"),
+	...require("./channel-flags.js"),
+	require("./discord.js"),
+	require("./gachi.js"),
+	require("./language.js"),
+	require("./location.js"),
+	require("./no-abb-chatter.js"),
+	require("./reminder.js"),
+	require("./suggestion.js"),
+	require("./timer.js"),
+	require("./track-favourite.js"),
+	...require("./twitch-lotto.js")
+];
